+++ modules/AWSPowerShell/AWSPowerShellCompleters.psm1
@@ -3117,7 +3117,7 @@ $AAR_Completers = {
             ($_ -eq "Update-AARService/SourceConfiguration_CodeRepository_CodeConfiguration_CodeConfigurationValues_Runtime")
         }
         {
-            $v = "CORRETTO_11","CORRETTO_8","NODEJS_12","NODEJS_14","PYTHON_3"
+            $v = "CORRETTO_11","CORRETTO_8","NODEJS_12","NODEJS_14","NODEJS_16","PYTHON_3"
             break
         }
 
@@ -31002,6 +31002,27 @@ $LS_Completers = {
             break
         }
 
+        # Amazon.Lightsail.HttpEndpoint
+        "Update-LSInstanceMetadataOption/HttpEndpoint"
+        {
+            $v = "disabled","enabled"
+            break
+        }
+
+        # Amazon.Lightsail.HttpProtocolIpv6
+        "Update-LSInstanceMetadataOption/HttpProtocolIpv6"
+        {
+            $v = "disabled","enabled"
+            break
+        }
+
+        # Amazon.Lightsail.HttpTokens
+        "Update-LSInstanceMetadataOption/HttpTokens"
+        {
+            $v = "optional","required"
+            break
+        }
+
         # Amazon.Lightsail.InstanceAccessProtocol
         "Get-LSInstanceAccessDetail/Protocol"
         {
@@ -31012,7 +31033,7 @@ $LS_Completers = {
         # Amazon.Lightsail.InstanceMetricName
         "Get-LSInstanceMetricData/MetricName"
         {
-            $v = "BurstCapacityPercentage","BurstCapacityTime","CPUUtilization","NetworkIn","NetworkOut","StatusCheckFailed","StatusCheckFailed_Instance","StatusCheckFailed_System"
+            $v = "BurstCapacityPercentage","BurstCapacityTime","CPUUtilization","MetadataNoToken","NetworkIn","NetworkOut","StatusCheckFailed","StatusCheckFailed_Instance","StatusCheckFailed_System"
             break
         }
 
@@ -31147,6 +31168,9 @@ $LS_map = @{
     "CacheBehaviorSettings_ForwardedHeaders_Option"=@("New-LSDistribution","Update-LSDistribution")
     "ComparisonOperator"=@("Add-LSAlarm")
     "DefaultCacheBehavior_Behavior"=@("New-LSDistribution","Update-LSDistribution")
+    "HttpEndpoint"=@("Update-LSInstanceMetadataOption")
+    "HttpProtocolIpv6"=@("Update-LSInstanceMetadataOption")
+    "HttpTokens"=@("Update-LSInstanceMetadataOption")
     "IpAddressType"=@("New-LSDistribution","New-LSInstance","New-LSInstancesFromSnapshot","New-LSLoadBalancer","Set-LSIpAddressType")
     "MetricName"=@("Add-LSAlarm","Get-LSBucketMetricData","Get-LSContainerServiceMetricData","Get-LSDistributionMetricData","Get-LSInstanceMetricData","Get-LSLoadBalancerMetricData","Get-LSRelationalDatabaseMetricData")
     "Origin_ProtocolPolicy"=@("New-LSDistribution","Update-LSDistribution")
@@ -31363,6 +31387,7 @@ $LS_SelectMap = @{
                "Update-LSDistribution",
                "Update-LSDistributionBundle",
                "Update-LSDomainEntry",
+               "Update-LSInstanceMetadataOption",
                "Update-LSLoadBalancerAttribute",
                "Update-LSRelationalDatabase",
                "Update-LSRelationalDatabaseParameter")
@@ -36911,7 +36936,7 @@ $NS_Completers = {
         # Amazon.NimbleStudio.StreamingInstanceType
         "New-NSStreamingSession/Ec2InstanceType"
         {
-            $v = "g4dn.12xlarge","g4dn.16xlarge","g4dn.2xlarge","g4dn.4xlarge","g4dn.8xlarge","g4dn.xlarge"
+            $v = "g3.4xlarge","g3s.xlarge","g4dn.12xlarge","g4dn.16xlarge","g4dn.2xlarge","g4dn.4xlarge","g4dn.8xlarge","g4dn.xlarge","g5.16xlarge","g5.2xlarge","g5.4xlarge","g5.8xlarge","g5.xlarge"
             break
         }
 
@@ -48309,7 +48334,7 @@ $SSM_Completers = {
         # Amazon.SimpleSystemsManagement.DocumentType
         "New-SSMDocument/DocumentType"
         {
-            $v = "ApplicationConfiguration","ApplicationConfigurationSchema","Automation","Automation.ChangeTemplate","ChangeCalendar","Command","DeploymentStrategy","Package","Policy","ProblemAnalysis","ProblemAnalysisTemplate","Session"
+            $v = "ApplicationConfiguration","ApplicationConfigurationSchema","Automation","Automation.ChangeTemplate","ChangeCalendar","CloudFormation","Command","ConformancePackTemplate","DeploymentStrategy","Package","Policy","ProblemAnalysis","ProblemAnalysisTemplate","Session"
             break
         }