+++ modules/AWSPowerShell/Cmdlets/Lightsail/Basic/Get-LSInstanceMetricData-Cmdlet.cs
@@ -118,7 +118,11 @@ namespace Amazon.PowerShell.Cmdlets.LS
         /// or failed the instance status check. This metric can be either 0 (passed) or 1 (failed).
         /// This metric data is available in 1-minute (60 seconds) granularity.</para><para><code>Statistics</code>: The most useful statistic is <code>Sum</code>.</para><para><code>Unit</code>: The published unit is <code>Count</code>.</para></li><li><para><b><code>StatusCheckFailed_System</code></b> - Reports whether the instance passed
         /// or failed the system status check. This metric can be either 0 (passed) or 1 (failed).
-        /// This metric data is available in 1-minute (60 seconds) granularity.</para><para><code>Statistics</code>: The most useful statistic is <code>Sum</code>.</para><para><code>Unit</code>: The published unit is <code>Count</code>.</para></li></ul>
+        /// This metric data is available in 1-minute (60 seconds) granularity.</para><para><code>Statistics</code>: The most useful statistic is <code>Sum</code>.</para><para><code>Unit</code>: The published unit is <code>Count</code>.</para></li><li><para><b><code>MetadataNoToken</code></b> - Reports the number of times that the instance
+        /// metadata service was successfully accessed without a token. This metric determines
+        /// if there are any processes accessing instance metadata by using Instance Metadata
+        /// Service Version 1, which doesn't use a token. If all requests use token-backed sessions,
+        /// such as Instance Metadata Service Version 2, then the value is 0.</para><para><code>Statistics</code>: The most useful statistic is <code>Sum</code>.</para><para><code>Unit</code>: The published unit is <code>Count</code>.</para></li></ul>
         /// </para>
         /// </summary>
         #if !MODULAR