+++ modules/AWSPowerShell/Cmdlets/WAFV2/Basic/New-WAF2IPSet-Cmdlet.cs
@@ -47,8 +47,9 @@ namespace Amazon.PowerShell.Cmdlets.WAF2
         /// <summary>
         /// <para>
         /// <para>Contains an array of strings that specifies zero or more IP addresses or blocks of
-        /// IP addresses in Classless Inter-Domain Routing (CIDR) notation. WAF supports all IPv4
-        /// and IPv6 CIDR ranges except for /0. </para><para>Example address strings: </para><ul><li><para>To configure WAF to allow, block, or count requests that originated from the IP address
+        /// IP addresses. All addresses must be specified using Classless Inter-Domain Routing
+        /// (CIDR) notation. WAF supports all IPv4 and IPv6 CIDR ranges except for <code>/0</code>.
+        /// </para><para>Example address strings: </para><ul><li><para>To configure WAF to allow, block, or count requests that originated from the IP address
         /// 192.0.2.44, specify <code>192.0.2.44/32</code>.</para></li><li><para>To configure WAF to allow, block, or count requests that originated from IP addresses
         /// from 192.0.2.0 to 192.0.2.255, specify <code>192.0.2.0/24</code>.</para></li><li><para>To configure WAF to allow, block, or count requests that originated from the IP address
         /// 1111:0000:0000:0000:0000:0000:0000:0111, specify <code>1111:0000:0000:0000:0000:0000:0000:0111/128</code>.</para></li><li><para>To configure WAF to allow, block, or count requests that originated from IP addresses